requirements.txt
streamlit
gtts
google-generativeai
python-dotenv
Pillow